+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add drain structure type with support in quick forms. [#316](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/316)
+
 ### Fixed
 
 - File validation error does not prevent from submission. [#318](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/318)